+++ priv/repo/migrations/20220611141322_create_assets.exs
@@ -5,7 +5,7 @@ defmodule AssetTracker.Repo.Migrations.CreateAssets do
     create table(:assets, primary_key: false) do
       add(:id, :uuid, primary_key: true)
       add(:name, :string, null: false)
-      add(:units, :float, null: false)
+      add(:units, :decimal, null: false)
       add(:user_id, references(:users, type: :uuid), null: false)
       add(:brokerage_id, references(:brokerages, type: :uuid), null: false)
       timestamps()
